DDP-4635 PR feedback * Extracted copy-pasted Jetty setup to separate classes * Extracted rate limit filter to separate class * Made rate limit values and preferred IP headers optional * Added logging so you can tell what IP header will be used at boot time
