+++ pepper-apis/src/main/java/org/broadinstitute/ddp/DataDonationPlatform.java
@@ -15,23 +15,13 @@ import static spark.Spark.port;
 import static spark.Spark.stop;
 import static spark.Spark.threadPool;
 
-import java.io.IOException;
 import java.time.Instant;
-import java.util.Collections;
-import java.util.Enumeration;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import javax.servlet.Filter;
-import javax.servlet.ServletContext;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.http.entity.ContentType;
 import org.broadinstitute.ddp.client.DsmClient;
 import org.broadinstitute.ddp.constants.ConfigFile;
@@ -48,16 +38,17 @@ import org.broadinstitute.ddp.db.StudyActivityDao;
 import org.broadinstitute.ddp.db.TransactionWrapper;
 import org.broadinstitute.ddp.db.UserDao;
 import org.broadinstitute.ddp.db.UserDaoFactory;
-import org.broadinstitute.ddp.exception.DDPException;
 import org.broadinstitute.ddp.filter.AddDDPAuthLoggingFilter;
 import org.broadinstitute.ddp.filter.DsmAuthFilter;
 import org.broadinstitute.ddp.filter.HttpHeaderMDCFilter;
 import org.broadinstitute.ddp.filter.MDCAttributeRemovalFilter;
 import org.broadinstitute.ddp.filter.MDCLogBreadCrumbFilter;
+import org.broadinstitute.ddp.filter.RateLimitFilter;
 import org.broadinstitute.ddp.filter.StudyLanguageContentLanguageSettingFilter;
 import org.broadinstitute.ddp.filter.StudyLanguageResolutionFilter;
 import org.broadinstitute.ddp.filter.TokenConverterFilter;
 import org.broadinstitute.ddp.filter.UserAuthCheckFilter;
+import org.broadinstitute.ddp.jetty.JettyConfig;
 import org.broadinstitute.ddp.json.errors.ApiError;
 import org.broadinstitute.ddp.model.dsm.DrugStore;
 import org.broadinstitute.ddp.monitoring.PointsReducerFactory;
@@ -154,31 +145,16 @@ import org.broadinstitute.ddp.util.LiquibaseUtil;
 import org.broadinstitute.ddp.util.LogbackConfigurationPrinter;
 import org.broadinstitute.ddp.util.ResponseUtil;
 import org.broadinstitute.ddp.util.RouteUtil;
-import org.eclipse.jetty.server.Server;
-import org.eclipse.jetty.server.session.SessionHandler;
-import org.eclipse.jetty.servlets.DoSFilter;
-import org.eclipse.jetty.util.thread.QueuedThreadPool;
-import org.eclipse.jetty.util.thread.ThreadPool;
 import org.quartz.Scheduler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.slf4j.MDC;
-import spark.ExceptionMapper;
 import spark.Request;
 import spark.Response;
 import spark.ResponseTransformer;
 import spark.Route;
 import spark.Spark;
-import spark.embeddedserver.EmbeddedServer;
-import spark.embeddedserver.EmbeddedServerFactory;
-import spark.embeddedserver.EmbeddedServers;
-import spark.embeddedserver.jetty.EmbeddedJettyServer;
-import spark.embeddedserver.jetty.HttpRequestWrapper;
-import spark.embeddedserver.jetty.JettyServerFactory;
-import spark.http.matching.MatcherFilter;
 import spark.route.HttpMethod;
-import spark.route.Routes;
-import spark.staticfiles.StaticFilesConfiguration;
 
 public class DataDonationPlatform {
 
@@ -244,9 +220,12 @@ public class DataDonationPlatform {
         // GAE specifies port to use via environment variable
         String appEnginePort = System.getenv(PORT);
 
-        EmbeddedServers.add(
-                EmbeddedServers.Identifiers.JETTY,
-                new AppEngineServerFactory(new JettyServer(), cfg.getString(ConfigFile.PREFERRED_SOURCE_IP_HEADER)));
+        String preferredSourceIPHeader = null;
+        if (cfg.hasPath(ConfigFile.PREFERRED_SOURCE_IP_HEADER)) {
+            preferredSourceIPHeader = cfg.getString(ConfigFile.PREFERRED_SOURCE_IP_HEADER);
+        }
+        JettyConfig.setupJetty(preferredSourceIPHeader);
+
         if (appEnginePort != null) {
             port(Integer.parseInt(appEnginePort));
         } else {
@@ -289,7 +268,7 @@ public class DataDonationPlatform {
         if (cfg.hasPath(ConfigFile.API_RATE_LIMIT.BURST)) {
             burst = cfg.getInt(ConfigFile.API_RATE_LIMIT.BURST);
         }
-        setupRateLimits(maxQueriesPerSecond, burst);
+        before("*", new RateLimitFilter(maxQueriesPerSecond, burst));
         before("*", new HttpHeaderMDCFilter(X_FORWARDED_FOR));
         before("*", new MDCLogBreadCrumbFilter());
 
@@ -535,23 +514,6 @@ public class DataDonationPlatform {
         LOG.info("ddp startup complete");
     }
 
-    private static void setupRateLimits(int maxRequestsPerSecond, int burst) {
-
-        DoSFilter rateLimitFilter = new DoSFilter();
-        rateLimitFilter.setEnabled(true);
-        try {
-            rateLimitFilter.init(new FilterConfig("DOSFilter", Map.of("delayMs", "-1",
-                    "maxRequestsPerSec", Integer.toString(maxRequestsPerSecond),
-                    "throttledRequests", Integer.toString(burst),
-                    "tooManyCode", "503")));
-        } catch (ServletException e) {
-            throw new DDPException("Could not initalize rate limits", e);
-        }
-        before("*", (req, res) -> {
-            rateLimitFilter.doFilter(req.raw(), res.raw(), (request, response) -> { });
-        });
-    }
-
     private static void setupApiActivityFilter() {
         afterAfter((Request request, Response response) -> {
             String endpoint = MDC.get(MDC_ROUTE_CLASS);
@@ -654,143 +616,4 @@ public class DataDonationPlatform {
         DBUtils.loadDaoSqlCommands(sqlConfig);
     }
 
-    /**
-     * RequestWrapper that uses a different header as the preferred source
-     * for the ip address that is returned by {@link HttpServletRequest#getRemoteAddr()}.
-     */
-    public static class SourceIPHeaderRequestWrapper extends HttpRequestWrapper {
-
-        String remoteAddr;
-
-        public SourceIPHeaderRequestWrapper(HttpServletRequest request, String preferredHeaderForIP) {
-            super(request);
-            String appEngineSourceIp = request.getHeader(preferredHeaderForIP);
-            remoteAddr = request.getRemoteAddr();
-            if (StringUtils.isNotBlank(appEngineSourceIp)) {
-                remoteAddr = appEngineSourceIp;
-            }
-        }
-
-        @Override
-        public String getRemoteAddr() {
-            return remoteAddr;
-        }
-    }
-
-    private static class AppEngineServerFactory implements EmbeddedServerFactory {
-        private final JettyServerFactory serverFactory;
-        private String preferredIPSourceHeader;
-        private ThreadPool threadPool;
-        private boolean httpOnly = true;
-
-        public AppEngineServerFactory(JettyServerFactory serverFactory, String preferredIPSourceHeader) {
-            this.serverFactory = serverFactory;
-            this.preferredIPSourceHeader = preferredIPSourceHeader;
-        }
-
-        public EmbeddedServer create(Routes routeMatcher, StaticFilesConfiguration staticFilesConfiguration,
-                                     ExceptionMapper exceptionMapper, boolean hasMultipleHandler) {
-            MatcherFilter matcherFilter = new MatcherFilter(routeMatcher, staticFilesConfiguration, exceptionMapper, false,
-                    hasMultipleHandler);
-            matcherFilter.init((FilterConfig)null);
-            SessionHandler handler = new JettyCustomRemoteAddrHeaderHandler(matcherFilter, this.preferredIPSourceHeader);
-            handler.getSessionCookieConfig().setHttpOnly(this.httpOnly);
-            return (new EmbeddedJettyServer(this.serverFactory, handler)).withThreadPool(this.threadPool);
-        }
-
-        public AppEngineServerFactory withThreadPool(ThreadPool threadPool) {
-            this.threadPool = threadPool;
-            return this;
-        }
-
-        public AppEngineServerFactory withHttpOnly(boolean httpOnly) {
-            this.httpOnly = httpOnly;
-            return this;
-        }
-    }
-
-    /**
-     * Lifted from {@link spark.embeddedserver.jetty.JettyHandler}, with one
-     * change to use {@link SourceIPHeaderRequestWrapper} to set
-     * the source IP properly in the appengine environment
-     */
-    public static class JettyCustomRemoteAddrHeaderHandler extends SessionHandler {
-        private javax.servlet.Filter filter;
-        private String preferredIPSourceHeader;
-
-        public JettyCustomRemoteAddrHeaderHandler(Filter filter, String preferredIPSourceHeader) {
-            this.filter = filter;
-            this.preferredIPSourceHeader = preferredIPSourceHeader;
-        }
-
-        public void doHandle(String target, org.eclipse.jetty.server.Request baseRequest, HttpServletRequest request,
-                             HttpServletResponse response) throws IOException, ServletException {
-            HttpRequestWrapper wrapper = new SourceIPHeaderRequestWrapper(request, preferredIPSourceHeader);
-            this.filter.doFilter(wrapper, response, null);
-            if (wrapper.notConsumed()) {
-                baseRequest.setHandled(false);
-            } else {
-                baseRequest.setHandled(true);
-            }
-        }
-    }
-
-    /**
-     * Lifted from {@link spark.embeddedserver.jetty.JettyServer}
-     */
-    public static class JettyServer implements JettyServerFactory {
-        JettyServer() {
-        }
-
-        public Server create(int maxThreads, int minThreads, int threadTimeoutMillis) {
-            Server server;
-            if (maxThreads > 0) {
-                int min = minThreads > 0 ? minThreads : 8;
-
-                int idleTimeout = threadTimeoutMillis > 0 ? threadTimeoutMillis : 30_000;
-                server = new Server(new QueuedThreadPool(maxThreads, min, idleTimeout));
-            } else {
-                server = new Server();
-            }
-
-            return server;
-        }
-
-        public Server create(ThreadPool threadPool) {
-            return threadPool != null ? new Server(threadPool) : new Server();
-        }
-    }
-
-    public static final class FilterConfig implements javax.servlet.FilterConfig {
-
-        private final String filterName;
-        private final Map<String, String> initParams;
-
-        public FilterConfig(String filterName,
-                            Map<String, String> initParams) {
-
-            this.filterName = filterName;
-            this.initParams = initParams;
-        }
-
-        @Override
-        public String getFilterName() {
-            return filterName;
-        }
-
-        @Override
-        public ServletContext getServletContext() {
-            return null;
-        }
-
-        @Override
-        public String getInitParameter(String s) {
-            return initParams.get(s);
-        }
-
-        @Override
-        public Enumeration<String> getInitParameterNames() {
-            return Collections.enumeration(initParams.keySet());
-        }
-    }
 }
